Move config model under model folder
