+++ apps/hub/src/run.ts
@@ -1,5 +1,5 @@
 import { Mqtt } from "@cornerstone/mqtt";
-import type { Config } from "./model.js";
+import type { Config } from "./model/config.js";
 
 export function run(config: Config) {
   const mqtt = new Mqtt(config.mqtt);
